refactor the turd that is makeLegend.js
