+++ css/webcharts.css
@@ -71,7 +71,7 @@
     }
 
     *[class*="wc-"] .legend {
-				width: 100%;
+        width: 100%;
         font-size: .9em;
         padding: 0;
         margin: 0;
